Tested changing proficiencies
Proficiencies can be chosen from a picker accessible through the '+' icon next to the header
They will be added to the 'list' (bindable layout), which will be automatically refreshed
If the user leaves without confirming changes or presses the delete icon next to it, the proficiency will be gone.
All other remaining properties basically work the same, except for actions and abilities, which will be text inputs. So it's just a matter of mostly copy-pasting